refactor: move cleansql to separate package
